requirements.txt
flask==3.0.0
python-dotenv==1.0.0
spotipy==2.23.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
gunicorn==20.1.0